+++ src/test/java/softeer2nd/chess/BoardTest.java
@@ -1,9 +1,9 @@
-package softeer2nd;
+package softeer2nd.chess;
 
-import org.junit.jupiter.api.*;
-import static org.junit.jupiter.api.Assertions.*;
+import org.junit.jupiter.api.Test;
+import softeer2nd.chess.pieces.Pawn;
 
-import softeer2nd.chess.domain.piece.Pawn;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class BoardTest {
     @Test
